+++ doc/whats_new.rst
@@ -50,6 +50,7 @@ Bug
   such as config and headshape are copied correctly, by `Mainak Jas`_ (`#135 <https://github.com/mne-tools/mne-bids/pull/135>`_)
 - Fix name of "sample" and "value" columns on events.tsv files, by `Ezequiel Mikulan`_ (`#185 <https://github.com/mne-tools/mne-bids/pull/185>`_)
 - Update function to copy KIT files to the `meg` directory, by `Matt Sanderson`_ (`#187 <https://github.com/mne-tools/mne-bids/pull/187>`_)
+- Allow raw data from KIT systems to have two marker files specified, by `Matt Sanderson`_ (`#167 <https://github.com/mne-tools/mne-bids/pull/173>`_)
 
 API
 ~~~